+++ src/scriptlets/set-cookie.js
@@ -38,7 +38,7 @@ import {
  *         - `accept`/ `accepted` / `notaccepted`
  *         - `reject` / `rejected`
  *         - `allow` / `allowed`
- *         - `disallow` / `deny`
+ *         - `disallow` / `deny` / `denied`
  *         - `enable` / `enabled`
  *         - `disable` / `disabled`
  *         - `necessary` / `required`